Dealing with cases where the http server returns gziped files
In some cases, the web server may be returning gziped content ranges.

This is problematic, because we require the requested offsets to be
relative to the uncompressed file, not the gziped version.

Correctly inspecting the responses requires the tools to know the
exact size of the source data for an index (not just the rounded
up block size).