Reducing a large amount of the cut-and-paste involved in patching files
The changes consolidate a lot of the higher level functionality into top level structures that do most of the work.
This makes it significantly easier to build upon and do things with the library, and hides a lot of the details involved in the most common use-case.

The first pass of this is a large change that also fixes some of the code to be more idiomatic in style.
There are some parts that have hard-coded assumptions that should be removed.